Fix for test case xc_misc.sql
A SELECT FOR UPDATE on a replicated table was being shipped by
mistake to all datanodes in a cluster with no primary node.
The problem was in function pgxc_FQS_find_datanodes.
The if condition to call GetPreferredReplicationNode
was only checking read relation access type, read for update was missing.
The result was that in case of read for update on a replicated table
in a cluster with no primary node, the query was being shipped to all
datanodes since the call to GetPreferredReplicationNode got skipped.
The fix was to add an additional check in the if condition.
In addition the patch adds assertions at places where GetRelationNodes
is called and a read for update on a replicated table is not expected.