Fix handling of number of tuples processed
Problem:
Currently the number of tuples processed is updated in
both HandleCommandComplete and ExecInsert/Update/Delete.

In HandleCommandComplete() it gets it from the command tag
returned from the datanode i.e. INSERT 0 2, UPDATE 5 and likewise.
And then it updates estate->es_processed. But it does this only
for FQS. For non-FQS, in ExecInsert/Update,  it is just incremented
by 1. So if a trigger function skips one row on datanode,
the command tag returned from datanode is INSERT 0 0.
But still in ExecInsert() increments the row count.

Solution:
Add a new field RemoteQueryState->rqs_processed, which is updated
in HandleCommandComplete(). Then it is used in ExecInsert/Update/Delete()
for non-FQS, and in RemoteQueryNext() for FQS.

Patch By Amit Khandekar
I just reviewed the patch, did a minor fix and committed the patch.