Regression test: select_views.sql
In select_views_2.out, masked the error "permission denied for relation"
for views. This is identified as a bug which could not be fixed easily.
Also, made ORDER BY related changes in the expected file and also in
alternate expected file
There are EXPLAIN statements in the test file. Because XC produces
different EXPLAIN output, changed the output of these statements in all
the expected files.