This patch fixes GTM bug to create more than 1024 sequences.
Due to hardcoded array size, GTM can not send to standby initial
list of the defined sequences.
So if GTM standby connects to GTM where more than 1024 sequences
defined, it causes memory corruption and segmentation fault.
The fix makes the array resizeable. For the same reason, client
is modified to accept sequence list of any size.

Andrei Martsinchyk reported this bug and provided the fix.