Another fix for the issue #259.   Wurenny reported and proposed the fix.
In XC, due to the scheduling of coordinator/GTM/datanode, GXID may not come
in order, or some GXID could be skipped in some nodes.   This happens very
often and the message level to notify this is LOG, which is always written
to log files.

The patch fixes this.

See https://github.com/postgres-x2/postgres-x2/issues/259 for details.
	modified:   src/backend/access/transam/varsup.c