In datanodes, the time of SQL execution is wrong
In coordinators, should contain the time of running in GTM, but in datanodes it should not be contained.
In fact, if you set the  parameter of "log_min_duration_statement" in the file of "postgres.conf", then you can see  the execution time of SQL. But you will find the time in datanodes is longer than in coordinators. It is obviously not true.